Crate tunneload

Tunneload is a Load-Balancer designed to run in a kubernetes Cluster and provide the needed Functionality to run, host and manage your web related Deployments without having to worry about routing or the like.

Modules

acceptors
Handles the way requests “enter” the load-balancer so via Webserver directly or tunneler for example

cli
Handles all the Command-Line stuff

configurator
All the Configuration stuff is handled by this module

forwarder
Handles the all things related to actually forwarding Requests to the Backend-Services

general
Contains some general helper functions that are not specific to one area but might be used in different parts of the project

handler
Handles already parsed requests and then also manages what should be done with a connection

htpasswd
Handles all the HTPasswd related stuff, which is mainly needed for the Basic-Auth middleware

metrics
Handles the Metrics related parts

rules
Handles all the Rule-Matching related stuff

tls
Contains some Generic functions and structs to establish and use TLS for an HTTPS connection

websockets
All the Websocket related functionality