Module tunneload::configurator

All the Configuration stuff is handled by this module

Modules

files

The File-Configurator allows users to use File(s) to configure the Tunneload Instance

kubernetes

The Kubernetes-Configurator allows users to dynamically configure the Tunneload Instance using different Ressource-Types in a Kubernetes-Cluster

parser

Contains all the relevant Parts for the Parser-Stuff

Structs

ConfigList

A List of different Types of Configurations

Manager

Manages all the Configuration for the Load-Balancer

ManagerBuilder

The Builder for creating a single Manager

RuleList

The List that contains all the Rules for Routing incoming Requests. This list is only used to update the Rules and not read from it.

Traits

ConfigItem

Defines an interface for a simple Configuration that can easily be managed

DefaultConfig

A Trait that allows Configs to generate a Default-Configuration using a given Name

Type Definitions

MiddlewareList

Holds a list of all Middlewares and provides a more ergonomic interface to interact with it

PluginList

The List of all Plugins currently registered

ServiceList

The List of all Services currently registered