Module tunneload::configurator::kubernetes

The Kubernetes-Configurator allows users to dynamically configure the Tunneload Instance using different Ressource-Types in a Kubernetes-Cluster

Modules

general

This handles all the Configuration using general Kubernetes Configurations

general_crd

Some general Functionality and Logic related to all CRDs

ingress

This is responsible for loading and parsing the Configurator using Kubernetes Ingress.

traefik_bindings

The Traefik specific stuff This loads Configuration from the Traefik Kubernetes CRDs

Structs

KubernetesConfigurator

The Dashboard-Entity for the Kubernetes Configurator

Functions

setup

Handles all the Setup related to the Kubernetes-Configurator