Module tunneload::configurator::kubernetes::general

This handles all the Configuration using general Kubernetes Configurations

Services

A Tunneload-Service is created for each Kubernetes Service with the same Name and Endpoints

TLS

Loads TLS Certifcates from Kubernetes tls secrets

Structs

KubernetesEvents

The Event-Emitter for the general Kubernetes-Configuration

KubernetesLoader

The Loader for the general Kubernetes-Configuration

KubernetesParser

This is the Parser for the general Kubernetes-Configuration

Functions

setup_general_configurator

Creates the General-Configurator for the Normal Kubernetes configurations with the given Namespace and kubernetes client