Module tunneload::configurator::kubernetes::ingress

This is responsible for loading and parsing the Configurator using Kubernetes Ingress.

Rules

Rules are created and updated based on Ingress Rules

Rule-Matchers

The Matchers for a Rule are simply obtained from the Ingress-HTTP spec which provides a Path, which will be used for a PathPrefix-Matcher, and a Domain, which will be used for a Domain-Matcher.

Rule-Service

The Service is simply obtained from the Target service of the Ingress-Rule and then searched for in Tunneloads internal list of Services

Rule-Middlewares

Middlewares for a Ingress-Rule are set using a custom annotation on the Ingress Object. To set a Middleware you simply need to add an Annotation with the tunneload-middleware Key and a comma seperated List of the Names of the Middlewares you want to use as the Value.

Rule-Priority

The Priority for a Rule is set to the Default Ingress Priority by default, however this can be overwritten on a by Rule bases by setting the “tunneload-priority” annotation with the desired Priority as the Value

Structs

IngressEvents

The Event-Emitter for the Kubernetes-Ingress-Configuration

IngressLoader

The Loader for the Kubernetes-Ingress-Configuration

IngressParser

The Parser for the Kubernetes-Ingress-Configuration

Functions

setup_general_configurator

Creates the General-Configurator for the Ingress configurations with the given Namespace, kubernetes client and defalt priority