Module tunneload::configurator::parser

Structs

GeneralConfigurator
The GeneralConfigurator is a general abstraction over a Loader and a Parser to allow for easier control over certain aspects of them as well as allowing for easier reuse of certain parts, like parsers, and better seperation of concerns

ParseRuleContext
The Context containing all the Data that could be needed for parsing a new Rule

RawMiddlewareConfig

RawRuleConfig

RawServiceConfig

RawTLSConfig

UnimplementedParserError
The Error returned by all the Default implementations for the Parser-Trait to indicate that the Parser does not implement the given Function

Enums

Event

MiddlewareParseError

Traits

EventEmitter

Loader

Parser

Functions

parse_middleware
Params:

Type Definitions

EventFuture