Trait tunneload::configurator::Configurator

pub trait Configurator {
#[must_use]    fn load_services<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Vec<Service>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn load_middleware<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Vec<Middleware>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn load_rules<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        middlewares: &'life1 MiddlewareList, 
        services: &'life2 ServiceList
    ) -> Pin<Box<dyn Future<Output = Vec<Rule>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn load_tls<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Vec<(String, CertifiedKey)>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_serivce_event_listener(
        &mut self, 
        services: ServiceList
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>;

    fn get_middleware_event_listener(
        &mut self, 
        middlewares: MiddlewareList
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>;

    fn get_rules_event_listener(
        &mut self, 
        middlewares: MiddlewareList, 
        services: ServiceList, 
        rules: RuleList
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>;

    fn get_tls_event_listener(
        &mut self, 
        tls_manager: ConfigManager
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>;
}

This Trait defines the Generic Interface for a Configurator Type

Required methods

#[must_use]fn load_services<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Vec<Service>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Loads all the Services from the Configurator

#[must_use]fn load_middleware<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Vec<Middleware>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Loads all the Middlewares from the Configurator

#[must_use]fn load_rules<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
middlewares: &'life1 MiddlewareList,
services: &'life2 ServiceList
) -> Pin<Box<dyn Future<Output = Vec<Rule>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Loads all the Rules from the Configurator

#[must_use]fn load_tls<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Vec<(String, CertifiedKey)>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Loads all the TLS-Configurations from the Configurator

fn get_serivce_event_listener(
&mut self,
services: ServiceList
) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>

Listens to all the Service related events and updates the Service configuration based on these events

fn get_middleware_event_listener(
&mut self,
middlewares: MiddlewareList
) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>

Listens to all the Middleware related events and updates the Middleware configuration based on these events

fn get_rules_event_listener(
&mut self,
middlewares: MiddlewareList,
services: ServiceList,
rules: RuleList
) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>

Listens to all the Rule related events and updates the Rule configuration based on these events

fn get_tls_event_listener(
&mut self,
tls_manager: ConfigManager
) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>

Listens to all the TLS-Configuration related events and updates the TLS-Configuration based on these events

Loading content...