Trait tunneload::configurator::DefaultConfig

pub trait DefaultConfig {
    fn default_name(name: String) -> Self;
}

A Trait that allows Configs to generate a Default-Configuration using a given Name

Usage:

This is needed for certain parts where the perceived Update-Order by the Load-Balancer can be out of order and therefore needs a way to create temporary Configs that effectively do nothing.

Required methods

fn default_name(name: String) -> Self

Returns a default Config with the given Name

Loading content...