Type Definition tunneload::configurator::PluginList

type PluginList = ConfigList<Plugin>;

The List of all Plugins currently registered

Implementations

impl PluginList

pub fn register_metrics(reg: &mut Registry)

This registers all the Prometheus Metrics related to service configuration

pub fn set_plugin_action(&self, n_srv: Plugin)

Inserts or Updates the given Service in the List of Services