Type Definition tunneload::configurator::ServiceList

type ServiceList = ConfigList<Service>;

The List of all Services currently registered

Implementations

impl ServiceList

pub fn register_metrics(reg: &mut Registry)

This registers all the Prometheus Metrics related to service configuration

pub fn set_service(&self, n_srv: Service)

Inserts or Updates the given Service in the List of Services