Trait tunneload::forwarder::Forwarder

pub trait Forwarder {
    type Connection: ServiceConnection;
    type ConnectError: Debug + Send;
    #[must_use]
    fn create_con<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        rule: &'life1 Rule
    ) -> Pin<Box<dyn Future<Output = Result<Self::Connection, Self::ConnectError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

A Forwarder is responsible for establishing a new Connection based on the provided Rule.

This Connection does not need to be an actual network connection, but rather can take any form that implements the ServiceConnection Trait.

Associated Types

type Connection: ServiceConnection

The underlying Type of the Connection that will be established by the Forwarder

type ConnectError: Debug + Send

The Error type returned by the Create_Con function

Required methods

#[must_use]
fn create_con<'life0, 'life1, 'async_trait>(
&'life0 self,
rule: &'life1 Rule
) -> Pin<Box<dyn Future<Output = Result<Self::Connection, Self::ConnectError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Attempts to create a new Connection based on the Data provided by the Rule