Trait tunneload::forwarder::ServiceConnection

pub trait ServiceConnection: Send + Sync + 'static {
    type WriteHalf: ServiceWriter;
    type ReadHalf: ServiceReader;
#[must_use]    fn read<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        buf: &'life1 mut [u8]
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn write<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        buf: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn halves_owned(self) -> (Self::ReadHalf, Self::WriteHalf);

#[must_use]    fn write_all<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        buf: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

#[must_use]    fn write_req<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        req: &'life1 Request<'life2>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Defines an Interface to send and receive Data between the Load-Balancer and a Service

This can take form in different ways and is not required to be done using an underlying networking connection, however this is most commonly the case.

Associated Types

type WriteHalf: ServiceWriter

type ReadHalf: ServiceReader

Loading content...
Required methods

#[must_use]fn read<'life0, 'life1, 'async_trait>(
&'life0 mut self,
buf: &'life1 mut [u8]
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Attempts to Read data from the Connection into the given Buffer-Array and returns how how many bytes were actually read and placed into the buffer

#[must_use]fn write<'life0, 'life1, 'async_trait>(
&'life0 mut self,
buf: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Attempts to Write the given Buffer to the underlying Connection and returns how many bytes were actually written

fn halves_owned(self) -> (Self::ReadHalf, Self::WriteHalf)

Splits the given Connection into an owned Read and Write halves

Loading content...
Provided methods

#[must_use]fn write_all<'life0, 'life1, 'async_trait>(
&'life0 mut self,
buf: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Writes the entire Buffer to the underlying connection

#[must_use]fn write_req<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
req: &'life1 Request<'life2>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Serializes the Request and then sends it over the underlying connection to the User

Loading content...