Module tunneload::internal_services

This module is responsible for dealing with and managing all the internal Tunneload Services, like the Dashboard

Modules

traits
Holds all the Traits needed for handling the internal stuff

Structs

ACMEHandler

Dashboard
This holds all the Information needed to provide the Tunneler-Dashboard as an internal Service

DashboardEntityList
This simply acts as a Wrapper to make it easier to manage Entities, especially in regards to serializing the Data into JSON or the like

Internals
Holds all the information regarding Internal-Services and Handlers, like the Dashboard and all the future services that may be included

Traits

DashboardEntity
The Bounds needed to register a new Entity on the Dashboard. This will provide all the Information that can then later be accessed and displayed on the Dashboard