Trait tunneload::internal_services::traits::InternalService

pub trait InternalService {
#[must_use]    fn handle<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self, 
        request: &'life1 Request<'life2>, 
        rule: Arc<Rule>, 
        sender: &'life3 mut dyn Sender
    ) -> Pin<Box<dyn Future<Output = Result<(), ()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait;

    fn check_service(&self, name: &str) -> bool;

    fn service(&self) -> Service;
}

This defines the Bounds for a single Internal Service

Required methods

#[must_use]fn handle<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
request: &'life1 Request<'life2>,
rule: Arc<Rule>,
sender: &'life3 mut dyn Sender
) -> Pin<Box<dyn Future<Output = Result<(), ()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

If a Request is bound for this Service, this function will be called with all the needed information to correctly handle this single Request

fn check_service(&self, name: &str) -> bool

Checks if the given Name matches up with the service’s internal Name, example “dashboard@internal”

fn service(&self) -> Service

The Service configuration that can then be targeted by Rules to access the Service

Loading content...