Trait tunneload::plugins::InstantiatePlugin

pub trait InstantiatePlugin {
    fn instantiate(
        name: String, 
        store: Store, 
        module: Module, 
        config: Option<Vec<u8>>
    ) -> Self;
}

This defines a simple interface to create actual Plugin-Instances, with config, from a basic loaded WASM-Plugin

Required methods

fn instantiate(
name: String,
store: Store,
module: Module,
config: Option<Vec<u8>>
) -> Self

Actually creates the new Instance of the Plugin with the given Configuration/data

Loading content...