Module tunneload::tls::auto

This handles all the parts related to Auto-TLS, which mostly revolves around obtaining and managing TLS-Certificates automatically so the end-user does not have to worry about it

Modules

discovery

This module contains all the given Discovery-Mechanisms that Tunneload provides by default

Structs

Account

A single Account used for generating and performing ACME-Challenges

AutoSession

Manages all the Auto-TLS-Session stuff

CertificateQueue

The Queue for Requested Certificates

CertificateRequest

A Request for a single new Certificate for the given Domain

ChallengeList

A Collection of all current ACME-Challenges

PendingTLS

This represents a single Pending ACME-Challenge

Enums

ChallengeState

This represents the individual states that a single Challenge can be in

Environment

The Let’s Encrypt Environment, this allows you to switch to the Staging environment when testing Configurations

NodeUpdateEvent

VerifyError

The Error type returned when generating Verifying Data

Traits

AutoDiscover

This Trait describes a basic interface that the Cluster relies upon to discover other Nodes that should also be moved into the Clustser

Functions

new

Creates all the Parts needed for the Automatic-TLS stuff

register_metrics

Registers all the related Metrics

renew

This is responsible for finding old certificates and then trying to renew them before they expire