Trait tunneload::tls::auto::AutoDiscover

pub trait AutoDiscover {
    fn get_own_id<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = NodeId> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_all_nodes<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = HashSet<NodeId>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn watch_nodes<'async_trait, D>(
        self: Arc<Self>, 
        raft: Arc<Cluster<D>>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        D: AutoDiscover + Send + Sync + 'static,
        D: 'async_trait,
        Self: 'async_trait;
}

This Trait describes a basic interface that the Cluster relies upon to discover other Nodes that should also be moved into the Clustser

Required methods

fn get_own_id<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = NodeId> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Gets the ID of the own/current Node

fn get_all_nodes<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = HashSet<NodeId>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Gets a Set of all currently known Nodes that are part of the Cluster

fn watch_nodes<'async_trait, D>(
self: Arc<Self>,
raft: Arc<Cluster<D>>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
D: AutoDiscover + Send + Sync + 'static,
D: 'async_trait,
Self: 'async_trait,

This should be the main task in which the Discovery-Mechanism runs and then updates the Cluster configuration as it discovers new Nodes for the Cluster