Trait tunneload::tls::auto::StoreTLS

pub trait StoreTLS {
#[must_use]    fn store_acc_key<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        priv_key: &'life1 PKey<Private>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn load_acc_key<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Option<PKey<Private>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn store<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        domain: String, 
        priv_key: &'life1 PKey<Private>, 
        certificate: &'life2 X509
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn private_key_to_bytes(key: &PKey<Private>) -> Option<Vec<u8>> { ... }

    fn cert_to_bytes(certificate: &X509) -> Option<Vec<u8>> { ... }
}

This defines a uniformi interface to allow for multiple Storage-Engines to be used to actually save the generated Certificates

Required methods

#[must_use]fn store_acc_key<'life0, 'life1, 'async_trait>(
&'life0 self,
priv_key: &'life1 PKey<Private>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

This is used to store the Private Key for the ACME-Account

#[must_use]fn load_acc_key<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Option<PKey<Private>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

This is used to load the Private Key for the ACME-Account

#[must_use]fn store<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
domain: String,
priv_key: &'life1 PKey<Private>,
certificate: &'life2 X509
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

This simply stores the single given Certificate for the Domain

Loading content...
Provided methods

fn private_key_to_bytes(key: &PKey<Private>) -> Option<Vec<u8>>

Turns the given Private-Key into the byte sequence that should be stored

fn cert_to_bytes(certificate: &X509) -> Option<Vec<u8>>

Turns the given Certificate into the byte sequence that should be stored

Loading content...