Trait tunneload::util::webserver::WebserverHandler

pub trait WebserverHandler {
    fn handle_request<'req, 'resp, 'life0, 'async_trait>(
        &'life0 self, 
        request: Request<'req>
    ) -> Pin<Box<dyn Future<Output = Result<Response<'resp>, ()>> + Send + 'async_trait>>
    where
        'req: 'resp,
        'req: 'async_trait,
        'resp: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;
}

This trait provides the general Interface the Webserver accepts its Handler to implement

Required methods

fn handle_request<'req, 'resp, 'life0, 'async_trait>(
&'life0 self,
request: Request<'req>
) -> Pin<Box<dyn Future<Output = Result<Response<'resp>, ()>> + Send + 'async_trait>> where
'req: 'resp,
'req: 'async_trait,
'resp: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

This function is called for every received Request.

Returns:

Ok(response): The Handler executed as planned and the given response should be send back to the user

Err(_): Some unrecoverable error occured in the Handler. The Server will then send an InternalServerError as response