Module tunneload::configurator::kubernetes::general

Some general functionality that is used by different Kubernetes parts

Structs

KubernetesEvents

KubernetesLoader

KubernetesParser

Watcher
A single custom Watcher that watches for any events related to the given Ressource-Type and ListParams

Enums

Event
The Events returned by the Watcher

Functions

get_tls_domain
Loads the Domain from the given TLS-Secret

load_secret
Attempts to load the Secret from the connected Kubernetes Cluster

parse_tls
Parses a givene Secret as a TLS-Secret and attempts to retrive the Domain and Key Pair to be used by the Acceptors for TLS