Trait tunneload::handler::traits::Handler

pub trait Handler: Debug {
    #[must_use]
    fn handle<'life0, 'async_trait, R, S>(
        &'life0 self, 
        id: u32, 
        receiver: R, 
        sender: S
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        R: Receiver + Send + 'static,
        S: Sender + Send + 'static,
        R: 'async_trait,
        S: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;
}

Defines a generic Handler that is responsible to handle incoming Connections, parse the Requests and route them to the right destinations as well as handling the responses

Required methods

#[must_use]
fn handle<'life0, 'async_trait, R, S>(
&'life0 self,
id: u32,
receiver: R,
sender: S
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
R: Receiver + Send + 'static,
S: Sender + Send + 'static,
R: 'async_trait,
S: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

Handles a single Connection